requirements.txt
python-liquid
python-frontmatter
pyyaml
markdown-it-py
Pillow